tutorial/requirement.txt
actionlib==1.12.0
amqp==1.4.9
angles==1.9.11
anyjson==0.3.3
billiard==3.3.0.23
bondpy==1.8.3
camera-calibration==1.13.0
camera-calibration-parsers==1.11.13
catkin==0.7.19
celery==3.1.18
controller-manager==0.15.1
controller-manager-msgs==0.15.1
cv-bridge==1.13.0
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
Django==2.2.7
djangorestframework==3.10.3
dynamic-reconfigure==1.6.0
gazebo-plugins==2.8.4
gazebo-ros==2.8.4
gencpp==0.6.2
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.12
gennodejs==2.0.1
genpy==0.6.9
image-geometry==1.13.0
interactive-markers==1.11.4
kdl-parser-py==1.13.1
kombu==3.0.37
laser-geometry==1.6.4
message-filters==1.14.3
pkg-resources==0.0.0
Pygments==2.5.1
python-qt-binding==0.3.6
pytz==2019.3
qt-dotgraph==0.3.15
qt-gui==0.3.15
qt-gui-cpp==0.3.15
qt-gui-py-common==0.3.15
redis==2.10.6
resource-retriever==1.12.5
rosbag==1.14.3
rosboost-cfg==1.14.7
rosclean==1.14.7
roscreate==1.14.7
rosgraph==1.14.3
roslaunch==1.14.3
roslib==1.14.7
roslint==0.11.2
roslz4==1.14.3
rosmake==1.14.7
rosmaster==1.14.3
rosmsg==1.14.3
rosnode==1.14.3
rosparam==1.14.3
rospy==1.14.3
rosservice==1.14.3
rostest==1.14.3
rostopic==1.14.3
rosunit==1.14.7
roswtf==1.14.3
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.8
rqt-dep==0.4.9
rqt-graph==0.4.10
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.13
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.9
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.5.1
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.9
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.6.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
rviz==1.13.5
sensor-msgs==1.12.7
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
sqlparse==0.3.0
tf==1.12.0
tf-conversions==1.12.0
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
topic-tools==1.14.3
urdfdom-py==0.4.1
xacro==1.13.4
